requirements.txt
flask
APScheduler
paho-mqtt
pymongo
firebase-admin
numpy
gpiozero
adafruit-circuitpython-dht
lgpio
RPi.GPIO